Refactor code to work with redux toolkit
