+++ src/App.js
@@ -1,24 +1,8 @@
-// // App.js
-// import React from 'react';
-// import Navigation from './Navigation';
-// import MainPage from './Mainpage';
-// import '../style.css';
-
-// function App() {
-//   return (
-//     <div className="main-container">
-//       <h1 className="rent-house">Rent a House</h1>
-//       <Navigation />
-//       <MainPage />
-//     </div>
-//   );
-// }
-
 import React from 'react';
 import { BrowserRouter as Routes, Route } from 'react-router-dom';
-import Navigation from './Navigation';
-import MainPage from './Mainpage';
-import HouseDetails from './HouseDetails';
+import Navigation from './components/sidebar/Navigation';
+import MainPage from './components/pages/Mainpage';
+import HouseDetails from './components/pages/HouseDetails';
 import '../style.css';
 
 function App() {
