+++ src/Redux/store.js
@@ -1,0 +1,8 @@
+import { configureStore } from '@reduxjs/toolkit';
+
+export const store = configureStore({
+  reducer: {},
+
+});
+
+export default store;
