+++ src/components/pages/HouseDetails.js
